@@ -14,7 +14,7 @@
 	<name>securingApp</name>
 	<description>securingApp</description>
 	<properties>
-		<java.version>10</java.version>
+		<java.version>11</java.version>
 	</properties>
 	<dependencies>
 		<dependency>
