@@ -6,6 +6,8 @@ import com.auth0.jwt.algorithms.Algorithm;
 import com.auth0.jwt.interfaces.DecodedJWT;
 import com.example.securingapp.sec.entities.AppRoles;
 import com.example.securingapp.sec.entities.AppUser;
+import com.example.securingapp.sec.security.JWTUtils;
+import com.example.securingapp.sec.security.RoleUsername;
 import com.example.securingapp.sec.service.AccountService;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.AllArgsConstructor;
@@ -18,6 +20,7 @@ import org.springframework.web.bind.annotation.RestController;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import java.security.Principal;
 import java.util.Date;
 import java.util.HashMap;
 import java.util.List;
@@ -44,42 +47,31 @@ public class accountRestController {
         return accountService.newAppRole(appRoles);
     }
     @PostMapping("/roleToUser")
-    public void addUser(@RequestBody roleUserForm r){
+    public void addUser(@RequestBody RoleUsername r){
          accountService.addRoleToUser(r.getRoleName() , r.getUserName());
     }
-    @GetMapping("/refreshToken")
+    @GetMapping(JWTUtils.REFRESH_TOKEN_ENDPOINT)
     public void refreshToken(HttpServletRequest request , HttpServletResponse response) throws Exception{
-        String authToken = request.getHeader("Authorization");
-        if(authToken!=null && authToken.startsWith("Bearer")){
+        String authToken = request.getHeader(JWTUtils.AUTH_HEADER);
+        if(authToken!=null && authToken.startsWith(JWTUtils.PREFIX)){
             try{
-                String jwt = authToken.substring(7);
-                Algorithm algorithm = Algorithm.HMAC256("secret!@#$");
-                JWTVerifier jwtVerifier = JWT.require(algorithm).build();
-                DecodedJWT decodedJWT = jwtVerifier.verify(jwt);
-                String username = decodedJWT.getSubject();
-                AppUser appUser = accountService.LoadUserByUserName(username);
-                //blacklist
-                String jwtAccessToken= JWT.create()
-                        .withSubject(appUser.getUsername())
-                        .withExpiresAt(new Date(System.currentTimeMillis()+2*60*1000))
-                        .withIssuer(request.getRequestURL().toString())
-                        .withClaim("role" , appUser.getAppRoles().stream().map(appRoles -> appRoles.getRoleName()).collect(Collectors.toList()))
-                        .sign(algorithm);
-                Map<String , String> idTokens = new HashMap<>();
-                idTokens.put("accessToken" , jwtAccessToken) ;
-                idTokens.put("refreshToken" , jwt) ;
-                response.setContentType("application/json");
-                new ObjectMapper().writeValue(response.getOutputStream() , idTokens);
-
+               String username = JWTUtils.verifyJwtRefreshToken(authToken);
+               AppUser appUser = accountService.LoadUserByUserName(username);
+                //blacklist ....
+                List<String> roles  = appUser.getAppRoles().stream().map(r -> r.getRoleName()).collect(Collectors.toList());
+                String jwtAccessToken= JWTUtils.generateAccessToken(request , appUser.getUsername() , roles );
+                JWTUtils.sendJwt(jwtAccessToken , authToken.substring(JWTUtils.PREFIX.length()) , response);
             }catch (Exception e){
                 response.setHeader("error-message" , e.getMessage());
                 response.sendError(HttpServletResponse.SC_FORBIDDEN);
             }
         }
+        else{
+            throw new RuntimeException("refresh token required");
+        }
     }
-}
-@Data
- class roleUserForm {
-    private String userName ;
-    private String roleName ;
+    @GetMapping("/profile")
+    public AppUser getProfile(Principal principal){
+        return accountService.LoadUserByUserName(principal.getName());
+    }
 }
