@@ -1,25 +1,17 @@
-package com.example.securingapp.sec.filters;
-
-
-
-import com.auth0.jwt.JWT;
-import com.auth0.jwt.algorithms.Algorithm;
-import com.fasterxml.jackson.databind.ObjectMapper;
+package com.example.securingapp.sec.security.filters;
+import com.example.securingapp.sec.security.JWTUtils;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.AuthenticationException;
 import org.springframework.security.core.userdetails.User;
 import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
-
 import javax.servlet.FilterChain;
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
-import java.util.Date;
-import java.util.HashMap;
-import java.util.Map;
+import java.util.List;
 import java.util.stream.Collectors;
 
 public class JwtAuthenticationFilter extends UsernamePasswordAuthenticationFilter {
@@ -40,29 +32,10 @@ public class JwtAuthenticationFilter extends UsernamePasswordAuthenticationFilte
     @Override
     protected void successfulAuthentication(HttpServletRequest request, HttpServletResponse response, FilterChain chain, Authentication authResult) throws IOException, ServletException {
         User user=(User) authResult.getPrincipal();
-        //generating JWT
-        Algorithm algorithm=Algorithm.HMAC256("secret!@#$");
-        String jwtAccessToken= JWT.create()
-                .withSubject(user.getUsername())
-                        .withExpiresAt(new Date(System.currentTimeMillis()+1*60*1000))
-                                .withIssuer(request.getRequestURL().toString())
-                                        .withClaim("role" , user.getAuthorities().stream().map(a -> a.getAuthority()).collect(Collectors.toList()))
-                                                .sign(algorithm);
-
-        String jwtRefreshToken = JWT.create()
-                .withSubject(user.getUsername())
-                .withExpiresAt(new Date(System.currentTimeMillis()+20*60*1000))
-                .withIssuer(request.getRequestURL().toString())
-                .sign(algorithm);
-        Map<String , String> idTokens = new HashMap<>();
-        idTokens.put("access-Token" , jwtAccessToken) ;
-        idTokens.put("refresh-Token" , jwtRefreshToken) ;
-        response.setContentType("application/json");
-        new ObjectMapper().writeValue(response.getOutputStream() , idTokens);
-
-        // sending Jwt
-
-//        response.setHeader("Authorization" , jwtAccessToken);
+        List<String> roles = user.getAuthorities().stream().map(a -> a.getAuthority()).collect(Collectors.toList());
+        String jwtAccessToken = JWTUtils.generateAccessToken(request , user.getUsername() , roles );
+        String jwtRefreshToken = JWTUtils.generateRefreshToken(request , user.getUsername() );
+        JWTUtils.sendJwt(jwtAccessToken , jwtRefreshToken , response);
     }
 
 
